Fixed PostgreSQLDatabse copy to stdout bug
This was caused by the database connection having a constructor that
doesn't connect tot the db. That was useful in the past as
performCopyToStdout created a new connection anayways and didn't use the
member one (as this was a pqxx connection). Not we can use the member
connection and so need to always construct it.