Reduced idle cpu usage and increased script response times
Instead of polling every file descriptor, checking for new data and then
waiting for one millisecond, we now use a single poll-syscall to check
for a change in any fd and wait for one second. This means we don't wait
doing nothing anymore while data is available and stop checking every ms
while nothing is available.